APIService/requirements.txt
llama-cpp-python==0.2.19
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2